dtype

+++ diffusion.py
@@ -38,7 +38,8 @@ class Diffusion(nn.Module):
 
         # create a noise array that we want to denoise
         x = torch.randn(n, 1, self.length, device=model.device)
-        l = (torch.tensor(labels, device=model.device) * 100).view(-1, 1)
+        l = (torch.tensor(labels, device=model.device,
+             dtype=torch.float32) * 100).view(-1, 1)
         # l = torch.ones(n, device=model.device) * label
 
         # loop through all timesteps
